requirements.txt
# NeuroRise Smart Alarm System Dependencies

# Core dependencies
numpy>=1.19.0
scipy>=1.5.0
Pillow>=8.0.0

# Machine Learning
xgboost>=1.4.0
scikit-learn>=0.24.0

# Signal Processing & Sleep Analysis
yasa>=0.5.0

# Hardware (Raspberry Pi)
RPi.GPIO>=0.7.0
adafruit-circuitpython-ssd1306>=2.12.0
adafruit-blinka>=6.0.0

# System
board>=1.0
busio>=5.0.0

# Optional for development
pytest>=6.0.0
black>=21.0.0